test: User storys
