+++ app/controllers/merchants_controller.rb
