+++ spec/examples.txt
@@ -1,15 +1,16 @@
 example_id                                         | status | run_time        |
 -------------------------------------------------- | ------ | --------------- |
-./spec/feature/merchant/dashboard_spec.rb[1:1:1:1] | passed | 0.18262 seconds |
-./spec/models/customer_spec.rb[1:1:1]              | passed | 0.00482 seconds |
-./spec/models/invoice_item_spec.rb[1:1:1]          | passed | 0.00295 seconds |
-./spec/models/invoice_item_spec.rb[1:1:2]          | passed | 0.00051 seconds |
-./spec/models/invoice_spec.rb[1:1:1]               | passed | 0.11914 seconds |
-./spec/models/invoice_spec.rb[1:1:2]               | passed | 0.00188 seconds |
-./spec/models/invoice_spec.rb[1:1:3]               | passed | 0.00178 seconds |
-./spec/models/invoice_spec.rb[1:1:4]               | passed | 0.00049 seconds |
-./spec/models/item_spec.rb[1:1:1]                  | passed | 0.00065 seconds |
-./spec/models/item_spec.rb[1:1:2]                  | passed | 0.0033 seconds  |
-./spec/models/item_spec.rb[1:1:3]                  | passed | 0.00035 seconds |
-./spec/models/merchant_spec.rb[1:1:1]              | passed | 0.02517 seconds |
-./spec/models/transaction_spec.rb[1:1:1]           | passed | 0.00321 seconds |
+./spec/feature/merchant/dashboard_spec.rb[1:1:1:1] | passed | 0.21515 seconds |
+./spec/feature/merchant/dashboard_spec.rb[1:1:1:2] | failed | 0.01782 seconds |
+./spec/models/customer_spec.rb[1:1:1]              | passed | 0.05687 seconds |
+./spec/models/invoice_item_spec.rb[1:1:1]          | passed | 0.00113 seconds |
+./spec/models/invoice_item_spec.rb[1:1:2]          | passed | 0.03604 seconds |
+./spec/models/invoice_spec.rb[1:1:1]               | passed | 0.00109 seconds |
+./spec/models/invoice_spec.rb[1:1:2]               | passed | 0.00059 seconds |
+./spec/models/invoice_spec.rb[1:1:3]               | passed | 0.00179 seconds |
+./spec/models/invoice_spec.rb[1:1:4]               | passed | 0.00653 seconds |
+./spec/models/item_spec.rb[1:1:1]                  | passed | 0.05956 seconds |
+./spec/models/item_spec.rb[1:1:2]                  | passed | 0.00225 seconds |
+./spec/models/item_spec.rb[1:1:3]                  | passed | 0.00112 seconds |
+./spec/models/merchant_spec.rb[1:1:1]              | passed | 0.03111 seconds |
+./spec/models/transaction_spec.rb[1:1:1]           | passed | 0.04214 seconds |
